Move Metrics and Traces into packages
